+++ packages/zip/index.d.ts
@@ -1,3 +1,1 @@
-export { readZip, writeZip } from './src/main.js';
-export { ZipReader } from './src/reader.js';
-export { ZipWriter } from './src/writer.js';
+export { ZipReader, readZip, ZipWriter, writeZip } from './index.js';
