+++ packages/zip/reader.js
@@ -1,0 +1,1 @@
+export { ZipReader } from './src/reader.js';
