+++ packages/zip/writer.js
@@ -1,0 +1,1 @@
+export { ZipWriter } from './src/writer.js';
